Fix compilation of usbip tools

+++ tools/usb/usbip/libsrc/vhci_driver.c
@@ -348,7 +348,7 @@ int usbip_vhci_get_free_port(uint32_t speed)
 	return -1;
 }
 
-int usbip_vhci_attach_device2(uint8_t port, int sockfd, uint32_t devid,
+int usbip_vhci_attach_device2(int port, int sockfd, uint32_t devid,
 		uint32_t speed) {
 	char buff[200]; /* what size should be ? */
 	char attach_attr_path[SYSFS_PATH_MAX];
@@ -382,7 +382,7 @@ static unsigned long get_devid(uint8_t busnum, uint8_t devnum)
 }
 
 /* will be removed */
-int usbip_vhci_attach_device(uint8_t port, int sockfd, uint8_t busnum,
+int usbip_vhci_attach_device(int port, int sockfd, uint8_t busnum,
 		uint8_t devnum, uint32_t speed)
 {
 	int devid = get_devid(busnum, devnum);
@@ -390,7 +390,7 @@ int usbip_vhci_attach_device(uint8_t port, int sockfd, uint8_t busnum,
 	return usbip_vhci_attach_device2(port, sockfd, devid, speed);
 }
 
-int usbip_vhci_detach_device(uint8_t port)
+int usbip_vhci_detach_device(int port)
 {
 	char detach_attr_path[SYSFS_PATH_MAX];
 	char attr_detach[] = "detach";
